Add a default command type, and make it parse out URLs by default.
The 'default' command runs when no specific command is given. The
primary purpose is to ensure that we can respond to non-command messages
in a few cases, or log information, or do _something_ in the background.

At the moment, it's noisy and prints a pointless message when it finds a
URL. Don't use in production at this commit.