Beginnings of refactoring the core.
main() is now more minimal. Spins up an IRC client, and a custom Actor
which will handle all message, and dispatch commands when identified.

The plan is to actorise almost everything possible. This helps in
isolating a lot of non-bot logic and ensures that it doesn't
accidentally get threaded through the entire bot.