Proof of concept works - adding command handlers.
Command handlers are structs that implement the `Actor: Send + Sync` trait. This
trait defines a common entry point, and enforces Send + Sync so in
theory structs implementing it can be sent between threads/tasks.

This allows for lots of Actors handling different message types
concurrently. The boilerplate is still quote onerous, but registration
of commands and handlers is now less fraught, and all the potential for
breakage is in the Actors. Handlers don't necessarily need to be actual
Actors either, but it's probably best that they are.